reorganaize user-settings.lspconfig
