+++ .nvim/lua/user-settings/nvim-lspconfig/servers/vue.lua
@@ -1,0 +1,33 @@
+local vue = {}
+
+vue.setup = function(nvim_lsp)
+  local my_capabilities_status, my_capabilities = pcall(require, "user-settings.nvim-lspconfig.capabilities")
+  if not my_capabilities_status then
+    return
+  end
+
+  local mason_lspconfig_status, mason_lspconfig = pcall(require, "mason-lspconfig")
+  if not mason_lspconfig_status then
+    return
+  end
+
+  mason_lspconfig.setup_handlers({
+    function()
+      -- Volar (Vue Language Server for Vue 3.x)
+      nvim_lsp.volar.setup({
+        capabilities = my_capabilities.capabilities,
+        -- use Take Over Mode
+        filetypes = {
+          "typescript",
+          "javascript",
+          "javascriptreact",
+          "typescriptreact",
+          "vue",
+          "json",
+        },
+      })
+    end,
+  })
+end
+
+return vue
